requirement.txt
# Core Dependencies
pymupdf==1.23.6             # PDF parsing (PyMuPDF)
python-docx==1.1.0          # Word document processing
transformers==4.30.0        # BERT/CLIP models
torch==2.0.1                # PyTorch framework

# Image Processing
Pillow==10.0.0              # Image loading

# Audio Processing
openai-whisper==20230918    # Speech-to-text

# Ubuntu
sudo apt-get install swig libfreetype6-dev

# macOS
brew install swig freetype

# Vector Processing
numpy==1.24.3               # Numerical computing
scikit-learn==1.3.0         # Dimensionality reduction

# API Server
fastapi==0.95.2             # Web framework
uvicorn==0.22.0             # ASGI server
python-multipart==0.0.6     # File upload support

# Utilities
langchain==0.0.247          # LLM orchestration
tqdm==4.65.0                # Progress bars